Numbers target animation closer to original
In the original Flemish version of Countdown, ie. "Cijfers&Letters",
each of the 3 digits was randomly chosen by pushing a button under the
digit.  The digits started spinning one after the other, ie. there was
no waiting for the previous digit to stop spinning before the next
started...  By making the `animation-delay` less than the overall
animation time, we achieve this effect.